versioncontrol: the way to get the youngest stored revision should depend on the version control back-end.
One can't assume that the sequence of revision is also ordered in time. 
This is especially true when using SVK for mirroring multiple repositories in one 
Subversion repository back-end: the timestamps of the mirrored changesets are kepts,
therefore the global ordering of changesets over time will not match the numerical
sequence of changesets.