Make it easier to remove the availability of the ''Ticket details'' event filter in the TracTimeline.
By default, it is disabled, but it can be restored by using a new config entry in TracIni:
{{{
[timeline]
ticket_details = yes
}}}

Also, even if the feature is enabled, the filter itself will be inactive by default
(i.e. the ''Ticket details'' check box is left unchecked; the user can of course
decide to use it).

In a future release, the `trac.ticket.UpdateDetailsForTimeline`
will be merged with the base ticket event provider, so don't
rely on the presence of that component to enable/disable the feature.

Fixes #1895.