Added the possibility to set an explicit id for headings.
e.g.
{{{
== My extra long heading subject to be modified == #but-keep-this-fixed-id
}}}

Also, note that ''id'' is an XML [http://www.w3.org/TR/REC-xml/#NT-Name ID], not an HTML id, as we're generating [http://www.w3.org/TR/xhtml1/#C_8 XHTML]. We were anyway not following strictly the HTML id spec, because of the presence of unicode characters (see #1224).

Note also that the above contradicts r3335 and 3406 w.r.t. the "." character in ids, but I realized since that the wiki syntax for referencing anchors only has trouble with an ending ".", so in most situations, it's safe to accept the ".".

Suggested by the maintainer of the Flumotion's FAQ.