Fix a bug in rendering CC: changes.
The comparison has to be done on the unobfuscated addresses, otherwise there's a possibility that no change will be detected in case there's a collision on the ''name@...'' part. This will have the undesirable effect that the change will be shown in the "changed from ''old'' to ''new''", with ''old'' and ''new'' lists  containing the full e-mail addresses.

Related to #153.