HTTP 1.1 support for TracStandalone:  add `--http11` option for server content using HTTP/1.1 protocol version.
Typically one can expect a speed-up by a factor of 4, due to the Keep-Alive (on small/medium pages; for huge pages like big changesets, Genshi will be the limiting factor and the time savings of using HTTP 1.1 won't be noticeable).

A couple of places which were missing the Content-Length header had to be fixed.

This also fixes the #8020 issue, which reportedly happens when IExplorer thinks it is talking to a 1.0 server.