#2647: improve the filter panel to better separate And/Or features.
The left side of the panel shows an ''And'' selector to add a filter to the current set of clauses, while the right side shows an ''Or'' button for adding a new set of filters.

This change also places the "[-]" delete constraint buttons on the left side, fixing #2405 on the way.