More robust rendering of the error page.
The catching of exception during evaluation of navigation contributors has been extended beyond !TracErrors, as a large range of errors can be triggered there, all likely to re-happen during rendering of the error page itself.

One category of such error are exceptions raised by a permission policy. Such an exception might be raised in any permission check, so we added a "second chance" rendering of the error page, with permission checks within the template disabled.

Closes #8585.