0.11.7dev: more straightforward regular expression for detecting CamelCase, taking directly into account the full list of upper and lower case unicode characters instead of playing tricks with negative look-behinds...
The previous support for unicode (#230) used a regexp which could lead to some pathological run-time in presence of unicode words that didn't contain any alphanumerical characters.

The runtime is now a tad bit slower and the WikiPageNames now also accept names like Page/Sub (which can be seen as a good thing), but the pathological behavior is gone.

Fixes #9025.