0.12.3dev: fix an unicode issue in the `[[Image]]` macro.
When given a project relative path (as in #10485), we shouldn't
produce an UTF-8 encoded `str` out of it, as this will trigger an
`UnicodeError` when mixed with `unicode` later on.

Modified all the Image macros to take some unicode input, to verify
the other code paths. Also removed the request parameters from the
image URL, in the `alt` and `title` attributes, to make them less
noisy.

Closes #10485.