1.0.2dev: second follow-up to r11689, no need for `check_catalog` param.
`get_available_locales()` is defined in translation.py and clearly
states that it corresponds to the list of locale identifiers for which
//translations// are available. Which really means that checking for
the presence of the compiled catalog is always the right thing to do
(modulo the 'en_US' exception).

The appropriate fix for `parse_date()` should have been to use the
list of locales for which we have //date localization information//
available, what the available translations are doesn't matter here.

That list corresponds to what Babel provides, so we use now for that
`get_known_locales` (= `babel.localedata.list`).