1.2dev: Owner change was not shown on ticket preview. Refs #11858.
The owner change was not shown for the special case of a 'new' ticket for which the component is being changed. Subject to some conditions, the owner is changed to the component owner in this case.

This issue was resolved by moving the //change owner to component owner for new tickets// logic to the `ConfigurableTicketWorkflow`.